A hack to permit to the Mruby side to receive the error code of the 'recv' system call.  This is needed to detect error #11 (EAGAIN) which is returned by recv when the socket is in non-block mode and there is no data to be received.
With this patch applied, in case of error, the two calls to recv
invoke mrb_sys_failreturn with a specifically string crafted in place
of the terse 'recv' string. The string is different for the two
instances, and includes both the numeric ID of the error and the
textual description. The string can be parsed from within ruby code
with this regexp:

/^(.*)\((\d*)\/(.*)\)/